Fix compilation for macOS and modern c++
When I try to compile, I get two errors, one due to macOS, which is that SOCK_NONBLOCK doesn't exist and instead they use O_NONBLOCK, and one due to c++, which is that there is now a `bind` function and so if you call the socket `bind()` while `using namespace std` there will be a conflict. With these changes I can build the project.